Move security submenu back to server page
